[test] Sample data: group all sample data under test.samples
