+++ composeApp/src/commonTest/kotlin/com/rwmobi/kunigami/data/source/network/AccountEndpointTest.kt
@@ -7,8 +7,8 @@
 
 package com.rwmobi.kunigami.data.source.network
 
-import com.rwmobi.kunigami.data.source.network.samples.GetAccountSampleData
 import com.rwmobi.kunigami.domain.exceptions.HttpException
+import com.rwmobi.kunigami.test.samples.GetAccountSampleData
 import io.ktor.client.HttpClient
 import io.ktor.client.engine.mock.MockEngine
 import io.ktor.client.engine.mock.respond
