+++ composeApp/src/commonTest/kotlin/com/rwmobi/kunigami/domain/usecase/GetConsumptionAndCostUseCaseTest.kt
@@ -11,7 +11,7 @@ import com.rwmobi.kunigami.domain.model.consumption.Consumption
 import com.rwmobi.kunigami.domain.model.consumption.ConsumptionTimeFrame
 import com.rwmobi.kunigami.domain.repository.FakeRestApiRepository
 import com.rwmobi.kunigami.domain.repository.FakeUserPreferencesRepository
-import com.rwmobi.kunigami.domain.samples.AccountSampleData
+import com.rwmobi.kunigami.test.samples.AccountSampleData
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.UnconfinedTestDispatcher
 import kotlinx.coroutines.test.runTest
