+++ composeApp/src/commonTest/kotlin/com/rwmobi/kunigami/domain/usecase/GetTariffSummaryUseCaseTest.kt
@@ -1,7 +1,7 @@
 package com.rwmobi.kunigami.domain.usecase
 
 import com.rwmobi.kunigami.domain.repository.FakeRestApiRepository
-import com.rwmobi.kunigami.domain.samples.TariffSampleData
+import com.rwmobi.kunigami.test.samples.TariffSampleData
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.UnconfinedTestDispatcher
 import kotlinx.coroutines.test.runTest
