+++ composeApp/src/commonTest/kotlin/com/rwmobi/kunigami/domain/usecase/InitialiseAccountUseCaseTest.kt
@@ -10,7 +10,7 @@ package com.rwmobi.kunigami.domain.usecase
 import com.rwmobi.kunigami.domain.exceptions.NoValidMeterException
 import com.rwmobi.kunigami.domain.repository.FakeRestApiRepository
 import com.rwmobi.kunigami.domain.repository.FakeUserPreferencesRepository
-import com.rwmobi.kunigami.domain.samples.AccountSampleData
+import com.rwmobi.kunigami.test.samples.AccountSampleData
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.UnconfinedTestDispatcher
 import kotlinx.coroutines.test.runTest
