+++ composeApp/src/commonTest/kotlin/com/rwmobi/kunigami/test/samples/AccountSampleData.kt
@@ -5,7 +5,7 @@
  *
  */
 
-package com.rwmobi.kunigami.domain.samples
+package com.rwmobi.kunigami.test.samples
 
 import com.rwmobi.kunigami.domain.model.account.Account
 import com.rwmobi.kunigami.domain.model.account.Agreement
