+++ composeApp/src/commonTest/kotlin/com/rwmobi/kunigami/test/samples/GetProductsSampleData.kt
@@ -5,7 +5,7 @@
  *
  */
 
-package com.rwmobi.kunigami.data.source.network.samples
+package com.rwmobi.kunigami.test.samples
 
 import com.rwmobi.kunigami.data.source.network.dto.LinkDto
 import com.rwmobi.kunigami.data.source.network.dto.products.ProductDetailsDto
